Address review comments
Rename dash_response to output_value, improve documentation,
don't make the jsonify call be in tail position.
